Provide hostname when setting cert thumbprint
This fixes the edge case where a node has been renamed and a new Puppet certificate was created for the node. Switching only the certificatethumbprint will fail as the hostname won't match the Subject Name of the new certificate. By also providing the fqdn fact to the Hostname parameter, switching becomes possible.